+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidyclust (development version)
+# tidyclust 0.1.2
 
 * The cluster specification methods for `generics::tune_args()` and `generics::tunable()` are now registered unconditionally (#115).